Don't fail unlink on ENOENT
Unlinking (deleting) the named pipe could error and fail if the file
didn't exist.

+++ index.ls
@@ -359,7 +359,7 @@ command = (line) ->
 input-stream = do
 
   mkfifo-stream = (path) ->
-    fs.unlink-sync path
+    if fs.exists-sync path then fs.unlink-sync path
     mkfifo-sync path, 8~600
     (spawn \tail [ \-F path ]).stdout
